requirements.txt
streamlit
numpy
plotly
sympy
pandas
Pillow
scipy
textblob 